ref #34: Create script for building worlds
Created a script that can generate worlds with a set amount of different
actors. It is also possible to specify if you want the actors shuffled
if you input multiple types.

An example of using the script to create 100 sprite actors and 200
physics actors in a shuffled order:
./create_world_with_n_actors shuffle 100 SpriteActors 200 PhysicsActors

Anoter example that creates 100 empty actors and 100 sprite actors in a
non shuffled order:
./create_world_with_n_actors 100 EmptyActors 100 SpriteActors