ref #34: Fix the output of the world creator
Fixed the output, as it was not valid json code that was outputed. It
seems that it is not as carefree as lua when it comes to certain things.
Also fixed an issue where the use of normal would treat all of the actor
instances as a single actor name in the script.