ref #71: Add the rest of the test runs
Added all the the remaining tests needed, both for gcc and clang.
Also removed the compilation test and the multi threaded test.
The compilation was removed as it became not needed, and the multi
threaded test is already covered in the exact same way in the numerious
unique components test.