fix #106: Finish ComponentCollection IndexMap
Finished the ComponentCollection implementation with the IndexMap,
allowing for log(n) searches.
The tradeoff however is that while the searches themselves are faster,
we have to do more searches.
Before doing an awake call was a linear search (within a defined search
space).
Now doing an awake is two binary searches (within the whole search
space), this is because have to keep
the map sorted at all times, rather than being able to sort it when
asked, as we don't know if find will be called in a non concurrent
environment.