fix #118: Null out non overloaded operations
Changed the functionality within the createMetaInformation function to
change the function pointers to nullptr if a function is not overloaded.
Meaning that the ComponentCollections can "quit" early.
I also made non overloaded functionality independent, so that we don't
have to spend a layer on each operation. However it would be better if
some of these operations were not scheduled at all.