fix #119: Fix seg fault issue
Fixed an issue with the execution layer algorithm that caused it to seg
fault. It was caused by the new feature allowing the algorithm to ignore
components that did not implement the specific function that was
considered in the algorithm. Previously this let me assume that all
components was always there in certain containers, since they never were
omitted, but now they are, and additional checks are needed to avoid
operating on non existant objects.