fix #130: Fix usage of reserved identifiers
The TypeIdentifier values in range 0-999 are reserved for NOX ECS
standard components. This has not been properly documented yet.

The test were using the type identifiers 0 and 1, which are reserved to
Children and Parent types components.
These components have special construction, as they are often
constructed through a factory, and adopted into a ComponentCollection,
rather than created within a ComponentCollection.

In this case they did not come through a factory, and therefore ended up
in a situation where they were not assigned an EntityManager pointer.