ref #108: Fix issue where tests crashed
Fixed the issue where the memory and fast spawning tests instantly
crashed. Problem was caused by the createMetainformation function that
did not have a return at the end of it, which was not for some reason an
error when compiling, probably because the compiler settings used by
nox.

fix #142
fix #143